rest-count/requirements.txt
Flask==0.10.1
python-consul==0.4.3
redis==2.10.3